Temperature Sensor RTD
1. The standard linear model for Pt100 RTD has a nominal resistance of 100 [Ω] at 0.0 [°C] and a sensitivity of 0.385 [Ω]
°C
a. Sketch a graph of the relationship between the resistance and temperature for Pt100 in temperature range from −40 [°C] to +120 [°C].
b. Find out what is the resistance value, when the temperature is 50 [°C]
c. What is the temperature, when the resistance is 120 [Ω]?
2. The second-order polynomial approximation for the resistance of Pt100 (having accuracy better than 0.01 [°C]) can be given by equation: 𝑅(𝑇) = 𝑅0(1 + 39.08 × 10−4 ⋅ 𝑇 + 5.8 × 10−7 ⋅ 𝑇2), where 𝑅0 is the nominal resistance.
a. b. c.
3. For a a. b. c.
Make tables and sketch graphs of the linear and second order models for Pt100 within temperature range from −40 [°C] to +120 [°C].
How much do the first order model differ from the second order approximation in the given temperature range?
What is the maximum difference in resistance and on which temperature does it occur?
selected NTC thermistor the reference resistance is 𝑅0 = 100 [𝑘Ω] and 𝛽 = 4600 [𝐾]. Sketch a graph of the resistance vs. temperature within a range from 0 [°C] to 100 [°C] Find the resistance values at temperature of 65 [°C].
What is the temperature, when the resistance of the NTC thermistors shows value of 20.5 [𝑘Ω]?
4. Below is an example circuit of LM35 for measuring temperatures from 2 to 150 [°C].
a. Sketch a graph of the output voltage vs. temperature within the given temperature
range.
b. What is the output voltage, when the temperature is 22 [°C]?
c. If the output voltage is 1.10 [𝑉], what has been the temperature?
5. Below is an example circuit of LM35 for measuring full-range temperatures from −55 [°C] to +150 [°C].
a. If the voltage source is 𝑉 = 5.0 [𝑉], what value for R1 should be chosen? 𝑠
b. Find the mathematical relationship between the output voltage and the temperature.
c. Sketch a graph of the relationship within the given measurement range.
d. If the output voltage is 650 [𝑚𝑉], what is then the temperature?